Refactor `field=` lookup behaviour
In the past, `field=…` was treated as `field__contains=…`. That became a
problem when #1041 fixed `__contains` and the behaviour became
confusingly unlike the Django ORM.

This patch introduces a new lookup type (`matches`) which is like
`__exact` in that it does not find values inside words (`field=food`
will not match "dogfood") but, to maintain some degree of backwards
compatibility, does not impose the strict ordering requirements of
`exact` (`field="foo bar"` is treated as `foo AND bar` with no
requirement that the values be adjacent or in that order).

The default operator for a simple `field=` lookup is now `matches`
rather than contains.